data_processing/requirements.txt
numpy==1.18.1
pillow==8.3.2
tqdm==4.62.2
matplotlib==3.3.4
scipy==1.5.4
click==8.0.4
